+++ src/webapp/src/components/analyze/Design.vue
@@ -276,10 +276,16 @@ export default {
 
       <div class="column">
         <div class="field is-grouped is-pulled-right">
-          <p v-if="hasActiveReport" class="control" @click="getDashboards">
+          <p
+            v-if="hasActiveReport"
+            class="control"
+            data-test-id="dropdown-add-to-dashboard"
+            @click="getDashboards"
+          >
             <Dropdown label="Add to Dashboard" is-right-aligned>
               <div class="dropdown-content">
                 <a
+                  data-test-id="button-new-dashboard"
                   class="dropdown-item"
                   data-dropdown-auto-close
                   @click="toggleNewDashboardModal()"
@@ -325,7 +331,11 @@ export default {
             </Dropdown>
           </p>
 
-          <div class="control field" :class="{ 'has-addons': hasActiveReport }">
+          <div
+            class="control field"
+            :class="{ 'has-addons': hasActiveReport }"
+            data-test-id="dropdown-save-report"
+          >
             <p class="control">
               <button
                 v-if="hasActiveReport"
@@ -363,6 +373,7 @@ export default {
                         Cancel
                       </button>
                       <button
+                        data-test-id="button-save-report"
                         class="button"
                         :disabled="!saveReportSettings.name"
                         data-dropdown-auto-close
@@ -442,6 +453,7 @@ export default {
                   <div class="field has-addons">
                     <div class="control">
                       <button
+                        data-test-id="run-query-button"
                         class="button is-success"
                         :class="{ 'is-loading': isLoadingQuery }"
                         :disabled="!currentSQL"
@@ -590,6 +602,7 @@ export default {
                   <a
                     v-if="!column.hidden"
                     :key="column.label"
+                    :data-test-id="`column-${column.label}`.toLowerCase()"
                     class="panel-block space-between has-text-weight-medium"
                     :class="{ 'is-active': column.selected }"
                     @click="columnSelected(column)"
@@ -628,6 +641,7 @@ export default {
                 <a
                   v-for="aggregate in design.relatedTable.aggregates"
                   :key="aggregate.label"
+                  :data-test-id="`aggregate-${aggregate.label}`.toLowerCase()"
                   class="panel-block space-between has-text-weight-medium"
                   :class="{ 'is-active': aggregate.selected }"
                   @click="aggregateSelected(aggregate)"