+++ src/webapp/src/components/model/AnalyzeModels.vue
@@ -128,7 +128,12 @@ export default {
               below.
             </p>
           </div>
-          <div v-for="(v, model) in models" :key="`${model}-panel`" class="box">
+          <div
+            v-for="(v, model) in models"
+            :key="`${model}-panel`"
+            :data-test-id="`${model}-model-card`.replace('/', '-')"
+            class="box"
+          >
             <div class="content">
               <div class="level level-tight">
                 <div class="level-left">