+++ src/webapp/src/components/pipelines/Extractors.vue
@@ -65,6 +65,7 @@ export default {
       <div
         v-for="(extractor, index) in plugins.extractors"
         :key="`${extractor}-${index}`"
+        :data-test-id="`${extractor}-extractor-card`"
         class="tile is-parent is-3 is-relative"
       >
         <SpeedRunIcon v-if="extractor === speedRunExtractor" />