+++ src/webapp/tests/e2e/specs/google-analytics.spec.js
@@ -1,6 +1,6 @@
 describe('Usage data tracking', function() {
   it('Checks for Google Analytics', function() {
-    cy.visit('http://localhost:5000/')
+    cy.visit('/')
     cy.window().then(win => {
       // GA is an acronym for Google Analytics
       const appGAId = Object.keys(win.gaData)[0]